add decimalsOffset = 3 + fixed tests, cause shares were messed with assets

+++ src/contracts/interfaces/IStakeToken.sol
@@ -5,8 +5,8 @@ import {IERC4626} from 'openzeppelin-contracts/contracts/interfaces/IERC4626.sol
 
 interface IStakeToken is IERC4626 {
   struct CooldownSnapshot {
-    /// @notice Amount of tokens available for withdrawal
-    uint192 amount;
+    /// @notice Amount of shares available to redeem
+    uint224 amount;
     /// @notice Time to unlock funds for withdrawal
     uint32 timestamp;
   }
